+++ packages/codeboltjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @codebolt/codeboltjs
 
+## 5.0.8
+
+### Patch Changes
+
+- updated the types and Structure for clientsdk
+- Updated dependencies
+  - @codebolt/types@5.0.8
+
 ## 5.0.7
 
 ### Patch Changes